Parameter assignment, draft 2
1. switched from assign_condition to assign_parameters. Assign
parameters picks a condition, then picks a subcondition, and (for yoked
conditions only) a yoking subject id and yoking sequence. It guarantees
that yoked conditions will only be chosen after adaptive varied (to
which they are yoked) and that each subject id in adaptive varied will
be used equally often as a yoking id.
2. modified demo.html to use assign_parameters.php and also to run
simulations of multiple subjects when needed
3. added schema list as global var to experiment.js, and added dummy TFR
questions as placeholders